Change to dchan repo with gh-pages branch

+++ scripts/deploy-docs.sh
@@ -1,6 +1,14 @@
 #!/bin/bash
 set -e # exit with nonzero exit code if anything fails
 
+if [ "${TRAVIS_PULL_REQUEST}" != "false" ]; then
+    exit 0
+fi
+
+if [ "${TRAVIS_BRANCH}" != "master" ]; then
+    exit 0
+fi
+
 # clear and re-create the out directory
 rm -rf out || exit 0;
 mkdir out;
@@ -14,6 +22,8 @@ cp dchan.pdf dchan.html dchan.txt out
 
 # go to the out directory and create a *new* Git repo
 cd out
+
+mv dchan.html index.html
 git init
 
 # inside this git repo we'll pretend to be a new user
@@ -31,4 +41,4 @@ git commit -m "Deploy to GitHub Pages"
 # repo's gh-pages branch. (All previous history on the gh-pages branch
 # will be lost, since we are overwriting it.) We redirect any output to
 # /dev/null to hide any sensitive credential data that might otherwise be exposed.
-git push --force --quiet "https://${GH_TOKEN}@${GH_REF}" master > /dev/null 2>&1
+git push --force --quiet "https://${GH_TOKEN}@${GH_REF}" master:gh-pages > /dev/null 2>&1
